+++ src/download.ts
@@ -9,7 +9,7 @@ import { getLogger } from './logger';
 /**
  * Minimum required Okteto CLI version.
  */
-export const minimum = '3.17.0';
+export const minimum = '3.19.0';
 
 /**
  * Gets the platform-specific installation path for the Okteto CLI binary.